Start using ts-loader to speed builds up
