+++ demos/storybook/tsconfig.json
@@ -7,14 +7,15 @@
         "esModuleInterop": true,
         "allowSyntheticDefaultImports": true,
         "strict": true,
+        "sourceMap": false,
         "forceConsistentCasingInFileNames": true,
         "module": "esnext",
         "moduleResolution": "node",
         "resolveJsonModule": true,
         "isolatedModules": true,
-        "noEmit": true,
+        "noEmit": false,
         "jsx": "react"
     },
     "include": ["src", "stories"],
-    "exclude": [ "node_modules" ]
+    "exclude": [ "node_modules", ".storybook" ]
 }
